+++ Modules/UserManagement/Http/Requests/UpdateMembers.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Modules\UserManagement\Http\Requests;
+
+use App\User;
+
+use Illuminate\Validation\Rule;
+use Illuminate\Foundation\Http\FormRequest;
+
+class UpdateMembers extends FormRequest
+{
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            'users' => [
+                'required',
+                'array',
+                Rule::in(User::select('id')->get()->pluck('id')),
+            ],
+            'role_admins' => [
+                'required',
+                'array',
+                Rule::in(User::select('id')->get()->pluck('id')),
+            ],
+        ];
+    }
+
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+}